fixes mappers to set entity fields properly, adds cors support, adds response messages

+++ src/main/java/com/example/ecommerce/repository/CreditCardRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CreditCardRepository extends JpaRepository<CreditCard, Long> {
-    Optional<CreditCard> findByCardNo(Integer cardNo);
+    Optional<CreditCard> findByCardNo(Long cardNo);
     List<CreditCard> findByUserId(Long userId);
 }
